Fix drop of correct database after cache computation
Before, the dogos database was always dropped when the cache had
finished calculating. Now, the gatos are dropped when the cats
have been calculated. If other species or cases are added,
those cases have to be added by hand for the drop of the database
because the name of the database is hardcoded for the moment.